Multiple bug fixes * Sub-classed events have unique identifiers now, so that a matcher for sub-classed event never match the base event * Server default queue changed to AutoClassQueue, and have a default priority to 400, makes it work for most events * quote default to True for escape, since it is safer * Session dead lock risk, also remove the container parameter to use an internal container * Security improve on URL encode issues * Support basic authorization * Other small bugs

+++ tests/testConnection.py
@@ -3,6 +3,7 @@ Created on 2015/6/29
 
 @author: hubo
 '''
+from __future__ import print_function
 import unittest
 from event.event import Event, withIndices
 from event.core import Scheduler, PollEvent, TimerEvent, SystemControlEvent, SystemControlLowPriorityEvent
@@ -411,6 +412,9 @@ class TestConnection(unittest.TestCase):
         self.scheduler.main()
         self.assertEqual(ret, b'ABABABABABABABABABAB')
     def testSelfConnectionUnix(self):
+        if not hasattr(socket, 'AF_UNIX'):
+            print('Skip UNIX socket test because not supported')
+            return
         try:
             os.remove('/var/run/unixsocktest.sock')
         except:
@@ -438,6 +442,9 @@ class TestConnection(unittest.TestCase):
         self.scheduler.main()
         self.assertEqual(ret, b'ABABABABABABABABABAB')
     def testSelfConnectionUnixDgram(self):
+        if not hasattr(socket, 'AF_UNIX'):
+            print('Skip UNIX socket test because not supported')
+            return
         try:
             os.remove('/var/run/unixsocktestudp1.sock')
         except:
